fix: Trail - 길이 타입 수정

+++ backend/src/main/java/com/hanssarang/backend/mountain/domain/Trail.java
@@ -18,7 +18,7 @@ import java.time.LocalTime;
 public class Trail extends BaseEntity {
 
     private String name;
-    private int length;
+    private double length;
     private LocalTime goingUpTime;
     private LocalTime goingDownTime;
     private String risk;
